+++ src/main/java/codesquad/codestagram/advice/GlobalModelAttributeAdvice.java
@@ -1,0 +1,19 @@
+package codesquad.codestagram.advice;
+
+import codesquad.codestagram.domain.User;
+import jakarta.servlet.http.HttpSession;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ModelAttribute;
+
+@ControllerAdvice
+public class GlobalModelAttributeAdvice {
+
+    @ModelAttribute
+    public void addLoginUser(Model model, HttpSession session) {
+        User user = (User) session.getAttribute("loginUser");
+        if(user != null) {
+            model.addAttribute("loginUser", user);
+        }
+    }
+}
